# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "租户信息管理系统后端 API 服务",
        "title": "租户信息管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/reports/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取仪表盘汇总数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "仪表盘数据",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.DashboardData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/reports/fees/composition": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围内的费用构成数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "费用构成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期 (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/repository.FeeComposition"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/reports/income": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围内的收入统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "收入统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期 (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "month",
                            "type"
                        ],
                        "type": "string",
                        "description": "分组方式",
                        "name": "groupBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.IncomeReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/reports/maintenance/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围内的维修统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "维修统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期 (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.MaintenanceReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/reports/occupancy": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前出租率统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "出租率统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期 (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.OccupancyReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/reports/tenants/ranking": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取缴费金额租户排行榜",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "报表统计"
                ],
                "summary": "租户排行",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期 (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/repository.TenantFeeRanking"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repository.FeeComposition": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "feeType": {
                    "type": "string"
                }
            }
        },
        "repository.IncomeByMonth": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "repository.MaintenanceStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "repository.MaintenanceStatusStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "repository.TenantFeeRanking": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "tenantId": {
                    "type": "integer"
                },
                "tenantName": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "service.DashboardData": {
            "type": "object",
            "properties": {
                "activeContracts": {
                    "type": "integer"
                },
                "occupancyRate": {
                    "type": "number"
                },
                "occupiedRooms": {
                    "type": "integer"
                },
                "pendingFees": {
                    "type": "integer"
                },
                "pendingMaintenance": {
                    "type": "integer"
                },
                "totalRooms": {
                    "type": "integer"
                },
                "totalTenants": {
                    "type": "integer"
                },
                "unpaidAmount": {
                    "type": "number"
                }
            }
        },
        "service.IncomeReport": {
            "type": "object",
            "properties": {
                "byMonth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.IncomeByMonth"
                    }
                },
                "byType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.FeeComposition"
                    }
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "service.MaintenanceReport": {
            "type": "object",
            "properties": {
                "byStatus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.MaintenanceStatusStats"
                    }
                },
                "byType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.MaintenanceStats"
                    }
                }
            }
        },
        "service.OccupancyReport": {
            "type": "object",
            "properties": {
                "occupancyRate": {
                    "type": "number"
                },
                "occupiedRooms": {
                    "type": "integer"
                },
                "totalRooms": {
                    "type": "integer"
                },
                "vacantRooms": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}